requirements.txt
asgiref==3.5.2
dj-database-url==1.0.0
Django==3.2.5
django-filter==22.1
django-on-heroku==1.1.2
djangorestframework==3.13.1
djangorestframework-jsonapi==5.0.0
djangorestframework-simplejwt==5.2.0
inflection==0.5.1
psycopg2-binary==2.9.3
PyJWT==2.4.0
pytz==2022.2.1
sqlparse==0.4.2
whitenoise==6.2.0
